pyproject.toml
[tool.poetry]
name = "test_progect"
version = "0.1.0"
description = ""
authors = ["goaho7 <goaho7@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.0"
djangorestframework = "^3.15.2"
django-environ = "^0.11.2"
psycopg2-binary = "^2.9.9"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
